refactor(tilemap_renderer): split webgl.rs into webgl_helpers submodule
webgl.rs grew past the 1500-line per-file budget after the depth /
blend-mode / mipmap work. Move everything that does not reference
WebGlBackend state out into a new webgl/webgl_helpers.rs submodule,
wired via mod_interface the same way as other layered modules in
the crate. The webgl module now owns only the backend struct and
the two renderer helpers (SpriteRenderer / MeshRenderer).

Moved to webgl_helpers:

 - ArrayBuffer<T>  GPU-side Vec
 - SpriteInstanceData / MeshInstanceData + AsBytes impls + size asserts
 - setup_sprite_batch_vao / setup_mesh_batch_vao (the latter now takes
   geometry buffers directly instead of \&GpuGeometry so it stays
   decoupled from the adapter struct)
 - GpuResources / GpuTexture / GpuSprite / GpuGeometry / GpuBatch
 - Loadable + source_to_loadable + resolve_loadable
 - apply_blend (was a WebGlBackend method; now a free function)
 - index_format / apply_texture_filter / topology_to_gl

Struct fields are now pub inside the helper module so the adapter
can still build them directly; GpuTexture.filter / .mipmap lost the
leading underscore since clippy flags pub_underscore_fields. Added
Default derive and #[must_use] on the new pub lookup methods to
keep the clippy warning count at or below the baseline.

Result: webgl.rs drops from 1909 to 1351 lines; new helpers file is
713 lines. No behavior change.
